--- backend/libtiramisu.so
@@ -1,1 +1,0 @@
-/home/stellar/src/tiramisu/build/libtiramisu.so
