--- backend/pytiramisu.hpp
@@ -1,10 +1,0 @@
-//  Copyright (c) 2019-2020 Christopher Taylor
-//
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-//
-#pragma once
-#ifndef __PYTIRAMISU_HPP__
-#define __PYTIRAMISU_HPP__
-
-#endif
