+++ phyfleaux/api/__init__.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import
-
 __license__ = """
 Copyright (c) 2020 R. Tohid
