--- phyfleaux/backend/physl/nodes.py
@@ -1,8 +1,0 @@
-from __future__ import absolute_import
-
-_license__ = """ 
-Copyright (c) 2020 R. Tohid
-
-Distributed under the Boost Software License, Version 1.0. (See accompanying
-file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-"""
